fix(opencv): use float64 dtype

+++ benchmarks/benchmark_opencv.py
@@ -1,12 +1,13 @@
 import cv2
 import utils
+import numpy as np
 
 
 def load_image_from_path(path: str):
     image = cv2.imread(path, cv2.IMREAD_COLOR)
     # use rgba like a normal library
     image = cv2.cvtColor(image, cv2.COLOR_BGR2RGBA)
-
+    image = image.astype(np.float64)
     return image
